[mandiant] Move Mandiant class to base.py
